+++ apps/decodex/src/orchestrator/execution_thread_archive.rs
@@ -1,355 +1,12 @@
-use std::collections::HashSet;
+mod candidates;
+mod dispatch;
+mod model;
 
-#[cfg(not(test))] use crate::agent;
-use crate::{
-	agent::{AppServerThreadArchiveOutcome, AppServerThreadArchiveRequest},
-	orchestrator::{
-		AppServerProcessEnv, AppServerRunResult, IssueRunPlan, Result, ServiceConfig, StateStore,
-		TERMINAL_GUARDED_RUN_STATUS, WorkflowDocument,
-	},
+#[cfg(test)]
+pub(super) use candidates::{
+	completed_issue_thread_archive_candidates, terminal_thread_archive_backlog_candidates,
 };
-
-#[derive(Clone)]
-pub(super) struct ThreadArchiveCandidate {
-	pub(super) issue_id: String,
-	pub(super) issue_identifier: String,
-	pub(super) run_id: String,
-	pub(super) attempt_number: i64,
-	pub(super) thread_id: String,
-	pub(super) sequence_number: i64,
-}
-
-struct ThreadArchiveCandidateSource<'a> {
-	run_id: &'a str,
-	issue_id: &'a str,
-	issue_identifier: &'a str,
-	attempt_number: i64,
-	thread_id: &'a str,
-	sequence_number: Option<i64>,
-}
-
-pub(super) fn archive_completed_issue_threads_best_effort(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	issue_run: &IssueRunPlan,
-	process_env: &AppServerProcessEnv,
-	transport: &str,
-	run_result: &AppServerRunResult,
-) {
-	let current = ThreadArchiveCandidate {
-		issue_id: issue_run.issue.id.clone(),
-		issue_identifier: issue_run.issue.identifier.clone(),
-		run_id: issue_run.run_id.clone(),
-		attempt_number: issue_run.attempt_number,
-		thread_id: run_result.thread_id.clone(),
-		sequence_number: run_result.event_count.saturating_add(1),
-	};
-
-	archive_issue_threads_best_effort(
-		project,
-		state_store,
-		&issue_run.issue.id,
-		&issue_run.issue.identifier,
-		process_env,
-		transport,
-		Some(current),
-	);
-}
-
-pub(super) fn reconcile_terminal_thread_archive_backlog_best_effort(
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-) {
-	let process_env = AppServerProcessEnv::default();
-	let transport = workflow.frontmatter().agent().transport();
-	let candidates =
-		match terminal_thread_archive_backlog_candidates(state_store, project.service_id()) {
-			Ok(candidates) => candidates,
-			Err(error) => {
-				tracing::warn!(
-					?error,
-					project_id = project.service_id(),
-					"Failed to list terminal thread archive backlog; skipping this archive reconciliation pass."
-				);
-
-				return;
-			},
-		};
-
-	for candidate in candidates {
-		archive_completed_issue_thread_best_effort(
-			project,
-			state_store,
-			&process_env,
-			transport,
-			&candidate,
-		);
-	}
-}
-
-#[cfg(test)]
-pub(super) fn completed_issue_thread_archive_candidates(
-	state_store: &StateStore,
-	issue_run: &IssueRunPlan,
-	run_result: &AppServerRunResult,
-) -> Result<Vec<ThreadArchiveCandidate>> {
-	issue_thread_archive_candidates(
-		state_store,
-		&issue_run.issue.id,
-		&issue_run.issue.identifier,
-		Some(ThreadArchiveCandidate {
-			issue_id: issue_run.issue.id.clone(),
-			issue_identifier: issue_run.issue.identifier.clone(),
-			run_id: issue_run.run_id.clone(),
-			attempt_number: issue_run.attempt_number,
-			thread_id: run_result.thread_id.clone(),
-			sequence_number: run_result.event_count.saturating_add(1),
-		}),
-	)
-}
-
-#[cfg(test)]
-pub(super) fn terminal_thread_archive_backlog_candidates(
-	state_store: &StateStore,
-	project_id: &str,
-) -> Result<Vec<ThreadArchiveCandidate>> {
-	terminal_thread_archive_backlog_candidates_inner(state_store, project_id)
-}
-
-fn archive_issue_threads_best_effort(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	issue_id: &str,
-	issue_identifier: &str,
-	process_env: &AppServerProcessEnv,
-	transport: &str,
-	current: Option<ThreadArchiveCandidate>,
-) {
-	let fallback_candidate = current.clone();
-	let candidates =
-		match issue_thread_archive_candidates(state_store, issue_id, issue_identifier, current) {
-			Ok(candidates) => candidates,
-			Err(error) => {
-				tracing::warn!(
-					?error,
-					project_id = project.service_id(),
-					issue_id,
-					issue = issue_identifier,
-					"Failed to list completed issue threads for archive; archiving current thread only."
-				);
-
-				fallback_candidate.into_iter().collect()
-			},
-		};
-
-	for candidate in candidates {
-		archive_completed_issue_thread_best_effort(
-			project,
-			state_store,
-			process_env,
-			transport,
-			&candidate,
-		);
-	}
-}
-
-fn issue_thread_archive_candidates(
-	state_store: &StateStore,
-	issue_id: &str,
-	issue_identifier: &str,
-	current: Option<ThreadArchiveCandidate>,
-) -> Result<Vec<ThreadArchiveCandidate>> {
-	let mut seen_thread_ids = HashSet::new();
-	let mut candidates = Vec::new();
-
-	if let Some(current) = current {
-		push_thread_archive_candidate(
-			state_store,
-			&mut candidates,
-			&mut seen_thread_ids,
-			ThreadArchiveCandidateSource {
-				run_id: &current.run_id,
-				issue_id: &current.issue_id,
-				issue_identifier: &current.issue_identifier,
-				attempt_number: current.attempt_number,
-				thread_id: &current.thread_id,
-				sequence_number: Some(current.sequence_number),
-			},
-		)?;
-	}
-
-	for attempt in state_store.list_run_attempts_for_issue(issue_id)? {
-		if !completed_issue_archive_attempt_status(attempt.status()) {
-			continue;
-		}
-
-		if let Some(thread_id) = attempt.thread_id() {
-			push_thread_archive_candidate(
-				state_store,
-				&mut candidates,
-				&mut seen_thread_ids,
-				ThreadArchiveCandidateSource {
-					run_id: attempt.run_id(),
-					issue_id: attempt.issue_id(),
-					issue_identifier,
-					attempt_number: attempt.attempt_number(),
-					thread_id,
-					sequence_number: None,
-				},
-			)?;
-		}
-	}
-
-	Ok(candidates)
-}
-
-#[cfg(not(test))]
-fn terminal_thread_archive_backlog_candidates(
-	state_store: &StateStore,
-	project_id: &str,
-) -> Result<Vec<ThreadArchiveCandidate>> {
-	terminal_thread_archive_backlog_candidates_inner(state_store, project_id)
-}
-
-fn terminal_thread_archive_backlog_candidates_inner(
-	state_store: &StateStore,
-	project_id: &str,
-) -> Result<Vec<ThreadArchiveCandidate>> {
-	let mut seen_thread_ids = HashSet::new();
-	let mut candidates = Vec::new();
-
-	for attempt in state_store.list_run_attempts_for_project(project_id)? {
-		if !completed_issue_archive_attempt_status(attempt.status()) {
-			continue;
-		}
-
-		if let Some(thread_id) = attempt.thread_id() {
-			push_thread_archive_candidate(
-				state_store,
-				&mut candidates,
-				&mut seen_thread_ids,
-				ThreadArchiveCandidateSource {
-					run_id: attempt.run_id(),
-					issue_id: attempt.issue_id(),
-					issue_identifier: attempt.issue_id(),
-					attempt_number: attempt.attempt_number(),
-					thread_id,
-					sequence_number: None,
-				},
-			)?;
-		}
-	}
-
-	Ok(candidates)
-}
-
-fn completed_issue_archive_attempt_status(status: &str) -> bool {
-	matches!(
-		status,
-		"succeeded" | "failed" | "interrupted" | "terminated" | TERMINAL_GUARDED_RUN_STATUS
-	)
-}
-
-fn push_thread_archive_candidate(
-	state_store: &StateStore,
-	candidates: &mut Vec<ThreadArchiveCandidate>,
-	seen_thread_ids: &mut HashSet<String>,
-	source: ThreadArchiveCandidateSource<'_>,
-) -> Result<()> {
-	if !seen_thread_ids.insert(source.thread_id.to_owned())
-		|| run_has_terminal_thread_archive_event(state_store, source.run_id)?
-	{
-		return Ok(());
-	}
-
-	candidates.push(ThreadArchiveCandidate {
-		issue_id: source.issue_id.to_owned(),
-		issue_identifier: source.issue_identifier.to_owned(),
-		run_id: source.run_id.to_owned(),
-		attempt_number: source.attempt_number,
-		thread_id: source.thread_id.to_owned(),
-		sequence_number: source
-			.sequence_number
-			.unwrap_or(state_store.event_count(source.run_id)?.saturating_add(1)),
-	});
-
-	Ok(())
-}
-
-fn run_has_terminal_thread_archive_event(state_store: &StateStore, run_id: &str) -> Result<bool> {
-	for event_type in ["thread/archive", "thread/archive/discarded"] {
-		if state_store.run_has_protocol_event(run_id, event_type)? {
-			return Ok(true);
-		}
-	}
-
-	Ok(false)
-}
-
-fn archive_completed_issue_thread_best_effort(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	process_env: &AppServerProcessEnv,
-	transport: &str,
-	candidate: &ThreadArchiveCandidate,
-) {
-	let archive_request = AppServerThreadArchiveRequest {
-		run_id: &candidate.run_id,
-		issue_id: &candidate.issue_id,
-		attempt_number: candidate.attempt_number,
-		listen: transport,
-		process_env,
-		thread_id: &candidate.thread_id,
-		sequence_number: candidate.sequence_number,
-	};
-	#[cfg(not(test))]
-	let archive_result = agent::archive_app_server_thread_after_success(&archive_request, state_store);
-	#[cfg(test)]
-	let archive_result = {
-		state_store
-			.append_event(
-				archive_request.run_id,
-				archive_request.sequence_number,
-				"thread/archive",
-				&serde_json::json!({
-					"threadId": archive_request.thread_id,
-					"issueId": archive_request.issue_id,
-					"attemptNumber": archive_request.attempt_number,
-				})
-				.to_string(),
-			)
-			.map(|()| AppServerThreadArchiveOutcome::Archived)
-	};
-
-	match archive_result {
-		Ok(AppServerThreadArchiveOutcome::Archived) => tracing::info!(
-			project_id = project.service_id(),
-			issue_id = candidate.issue_id,
-			issue = candidate.issue_identifier,
-			run_id = candidate.run_id,
-			attempt = candidate.attempt_number,
-			thread_id = %candidate.thread_id,
-			"Archived completed issue app-server thread."
-		),
-		Ok(AppServerThreadArchiveOutcome::DiscardedMissingThread) => tracing::info!(
-			project_id = project.service_id(),
-			issue_id = candidate.issue_id,
-			issue = candidate.issue_identifier,
-			run_id = candidate.run_id,
-			attempt = candidate.attempt_number,
-			thread_id = %candidate.thread_id,
-			"Discarded completed issue app-server thread archive because the thread is missing."
-		),
-		Err(error) => tracing::warn!(
-			?error,
-			project_id = project.service_id(),
-			issue_id = candidate.issue_id,
-			issue = candidate.issue_identifier,
-			run_id = candidate.run_id,
-			attempt = candidate.attempt_number,
-			thread_id = %candidate.thread_id,
-			"Failed to archive completed issue app-server thread; leaving completed run intact."
-		),
-	}
-}
+pub(super) use dispatch::{
+	archive_completed_issue_threads_best_effort,
+	reconcile_terminal_thread_archive_backlog_best_effort,
+};
