+++ apps/decodex/src/recovery/review_handoff.rs
@@ -1,11 +1,15 @@
 //! Review handoff recovery orchestration and validation.
 
+pub(in crate::recovery) mod commands;
+
 mod issue;
 mod labels;
+mod model;
 mod worktree;
 
 pub(super) use self::{
 	issue::load_issue_by_identifier,
+	model::{AdoptValidation, RebindLabelValidation, RebindValidation},
 	worktree::{relative_worktree_path_for_recovery, validate_retained_pr_worktree},
 };
 #[cfg(test)]
@@ -13,218 +17,20 @@ pub(super) use self::{
 	issue::validate_existing_handoff_refresh, worktree::validate_adopt_existing_worktree_mapping,
 };
 
-use std::{
-	env,
-	path::{Path, PathBuf},
-};
+use std::env;
 
 use crate::{
 	prelude::{Result, eyre},
-	pull_request::PullRequestLandingState,
 	recovery::{
-		context::{self, RecoveryContext},
+		context::RecoveryContext,
 		events,
 		git_worktree::{self},
 		pull_request_inspection::{self},
-		reports,
-		reports::ReviewHandoffRecoveryReport,
-		requests::{
-			ReviewHandoffAdoptRequest, ReviewHandoffDiagnoseRequest, ReviewHandoffRebindRequest,
-		},
-		review_handoff_apply,
-		review_handoff_diagnosis::{self},
-		review_handoff_policy::{self, RebindMode, RebindSuccessStateTransition},
+		requests::{ReviewHandoffAdoptRequest, ReviewHandoffRebindRequest},
+		review_handoff_policy::{self, RebindSuccessStateTransition},
 	},
-	state::WorktreeMapping,
-	tracker::{self, TrackerIssue},
+	tracker::TrackerIssue,
 };
-
-pub(super) struct RebindValidation {
-	pub(super) issue: TrackerIssue,
-	pub(super) worktree: WorktreeMapping,
-	pub(super) run_id: String,
-	pub(super) attempt_number: i64,
-	pub(super) landing_state: PullRequestLandingState,
-	pub(super) local_head_oid: String,
-	pub(super) worktree_path_for_event: Option<String>,
-	pub(super) active_label_present: bool,
-	pub(super) restore_active_label: bool,
-	pub(super) mode: RebindMode,
-	pub(super) success_state_transition: Option<RebindSuccessStateTransition>,
-	pub(super) clear_needs_attention_label: bool,
-}
-impl RebindValidation {
-	pub(super) fn should_restore_active_label(&self) -> bool {
-		self.restore_active_label
-	}
-}
-
-pub(super) struct AdoptValidation {
-	pub(super) issue: TrackerIssue,
-	pub(super) branch_name: String,
-	pub(super) worktree_path: PathBuf,
-	pub(super) run_id: String,
-	pub(super) attempt_number: i64,
-	pub(super) landing_state: PullRequestLandingState,
-	pub(super) local_head_oid: String,
-	pub(super) worktree_path_for_event: Option<String>,
-	pub(super) active_label_present: bool,
-	pub(super) success_state_transition: Option<RebindSuccessStateTransition>,
-	pub(super) previous_worktree_mapping: Option<WorktreeMapping>,
-}
-impl AdoptValidation {
-	pub(super) fn should_restore_active_label(&self) -> bool {
-		!self.active_label_present
-	}
-}
-
-pub(super) struct RebindLabelValidation {
-	pub(super) active_label_present: bool,
-	pub(super) restore_active_label: bool,
-	pub(super) clear_needs_attention_label: bool,
-}
-
-/// Run a read-only retained review handoff diagnostic.
-pub(crate) fn run_review_handoff_diagnose(
-	config_path: Option<&Path>,
-	request: &ReviewHandoffDiagnoseRequest,
-) -> Result<()> {
-	let context = context::load_recovery_context_read_only(config_path)?;
-
-	if let Some(message) = context::active_recovery_tracker_backoff_message(&context)? {
-		println!("{message}");
-
-		return Ok(());
-	}
-
-	let diagnostics = match match request.issue.as_deref() {
-		Some(issue_identifier) =>
-			review_handoff_diagnosis::diagnose_issue(&context, issue_identifier)
-				.map(|diagnostic| vec![diagnostic]),
-		None => review_handoff_diagnosis::diagnose_all_retained_review_worktrees(&context),
-	} {
-		Ok(diagnostics) => diagnostics,
-		Err(error) => {
-			if let Some(message) = context::remember_recovery_tracker_backoff_message(
-				&context,
-				&error,
-				"review_handoff_recovery",
-			) {
-				println!("{message}");
-
-				return Ok(());
-			}
-
-			return Err(error);
-		},
-	};
-	let report = ReviewHandoffRecoveryReport {
-		project_id: context.config.service_id().to_owned(),
-		diagnostics,
-	};
-
-	if request.json {
-		println!("{}", serde_json::to_string_pretty(&report)?);
-	} else {
-		print!("{}", reports::render_review_handoff_recovery_report(&report));
-	}
-
-	Ok(())
-}
-
-/// Run an explicit retained review handoff rebind.
-pub(crate) fn run_review_handoff_rebind(
-	config_path: Option<&Path>,
-	request: &ReviewHandoffRebindRequest,
-) -> Result<()> {
-	let context = context::load_recovery_context_for_dry_run(config_path, request.dry_run)?;
-	let validation = validate_rebind_request(&context, request)?;
-
-	if request.dry_run {
-		let state_transition = validation
-			.success_state_transition
-			.as_ref()
-			.map_or("none", |transition| transition.state_name.as_str());
-
-		println!(
-			"dry run: review handoff rebind validated for project={} issue={} branch={} pr={} head={} mode={} active_label_present={} would_restore_active_label={} state_transition={}",
-			context.config.service_id(),
-			validation.issue.identifier,
-			validation.worktree.branch_name(),
-			pull_request_inspection::landing_url(&validation.landing_state),
-			validation.local_head_oid,
-			validation.mode.evidence_value(),
-			validation.active_label_present,
-			validation.should_restore_active_label(),
-			state_transition
-		);
-
-		return Ok(());
-	}
-
-	review_handoff_apply::apply_review_handoff_rebind(&context, &validation)?;
-
-	println!(
-		"rebind ok: project={} issue={} branch={} pr={} head={} mode={}",
-		context.config.service_id(),
-		validation.issue.identifier,
-		validation.worktree.branch_name(),
-		pull_request_inspection::landing_url(&validation.landing_state),
-		validation.local_head_oid,
-		validation.mode.evidence_value()
-	);
-
-	Ok(())
-}
-
-/// Run an explicit manual PR takeover into retained review handoff state.
-pub(crate) fn run_review_handoff_adopt(
-	config_path: Option<&Path>,
-	request: &ReviewHandoffAdoptRequest,
-) -> Result<()> {
-	let context = context::load_recovery_context_for_dry_run(config_path, request.dry_run)?;
-	let validation = validate_adopt_request(&context, request)?;
-
-	if request.dry_run {
-		let state_transition = validation
-			.success_state_transition
-			.as_ref()
-			.map_or("none", |transition| transition.state_name.as_str());
-		let active_label = tracker::automation_active_label(context.config.service_id());
-
-		println!(
-			"dry run: review handoff adopt validated for project={} issue={} branch={} pr={} head={} run_id={} attempt={} active_label={} active_label_present={} would_restore_active_label={} state_transition={}",
-			context.config.service_id(),
-			validation.issue.identifier,
-			validation.branch_name,
-			pull_request_inspection::landing_url(&validation.landing_state),
-			validation.local_head_oid,
-			validation.run_id,
-			validation.attempt_number,
-			active_label,
-			validation.active_label_present,
-			validation.should_restore_active_label(),
-			state_transition
-		);
-
-		return Ok(());
-	}
-
-	review_handoff_apply::apply_review_handoff_adopt(&context, &validation)?;
-
-	println!(
-		"adopt ok: project={} issue={} branch={} pr={} head={} run_id={} attempt={}",
-		context.config.service_id(),
-		validation.issue.identifier,
-		validation.branch_name,
-		pull_request_inspection::landing_url(&validation.landing_state),
-		validation.local_head_oid,
-		validation.run_id,
-		validation.attempt_number
-	);
-
-	Ok(())
-}
 
 fn validate_rebind_request(
 	context: &RecoveryContext,
