+++ apps/decodex/src/recovery/review_handoff_diagnosis/binding/marker.rs
@@ -1,0 +1,62 @@
+use crate::{
+	recovery::{
+		git_worktree::{self, ReviewHandoffLineage},
+		review_handoff_diagnosis::binding::{
+			diagnostics,
+			model::{HandoffBindingDiagnostic, HandoffDiagnosticContext},
+		},
+	},
+	state::ReviewOrchestrationMarker,
+};
+
+pub(in crate::recovery) fn marker_head_binding_mismatch(
+	context: &HandoffDiagnosticContext<'_>,
+	local_head_oid: &str,
+	pr_base_ref: &Option<String>,
+	pr_head_oid: &Option<String>,
+) -> Option<HandoffBindingDiagnostic> {
+	let mismatch = if context.existing_handoff.pr_head_oid() != local_head_oid {
+		match git_worktree::worktree_head_descends_from_review_handoff(
+			context.worktree.worktree_path(),
+			context.existing_handoff.pr_head_oid(),
+			local_head_oid,
+		) {
+			ReviewHandoffLineage::Descends => None,
+			ReviewHandoffLineage::Diverged =>
+				Some(("review_handoff_lineage_mismatch", "review_handoff.pr_head_oid")),
+			ReviewHandoffLineage::Unknown =>
+				Some(("review_handoff_lineage_check_failed", "review_handoff.pr_head_oid")),
+		}
+	} else if let Some(orchestration) = context.existing_orchestration {
+		orchestration_binding_mismatch(context, orchestration, local_head_oid)
+	} else {
+		None
+	};
+
+	mismatch.map(|(reason, field)| {
+		diagnostics::rebind_required_diagnostic(
+			reason,
+			field,
+			pr_base_ref.clone(),
+			pr_head_oid.clone(),
+			context.issue_identifier,
+			context.existing_handoff.pr_url(),
+		)
+	})
+}
+
+fn orchestration_binding_mismatch(
+	context: &HandoffDiagnosticContext<'_>,
+	orchestration: &ReviewOrchestrationMarker,
+	local_head_oid: &str,
+) -> Option<(&'static str, &'static str)> {
+	if orchestration.branch_name() != context.worktree.branch_name() {
+		Some(("review_orchestration_branch_mismatch", "review_orchestration.branch_name"))
+	} else if orchestration.pr_url() != context.existing_handoff.pr_url() {
+		Some(("review_orchestration_pr_mismatch", "review_orchestration.pr_url"))
+	} else if orchestration.head_sha() != local_head_oid {
+		Some(("review_orchestration_head_mismatch", "review_orchestration.head_sha"))
+	} else {
+		None
+	}
+}
